Trezor support for Zcoin
