+++ lib/plugins.py
@@ -312,6 +312,8 @@ class DeviceMgr(ThreadJob, PrintError):
         # What we recognise.  Each entry is a (vendor_id, product_id)
         # pair.
         self.recognised_hardware = set()
+        # Custom enumerate functions for devices we don't know about.
+        self.enumerate_func = set()
         # For synchronization
         self.lock = threading.RLock()
         self.hid_lock = threading.RLock()
@@ -333,6 +335,9 @@ class DeviceMgr(ThreadJob, PrintError):
     def register_devices(self, device_pairs):
         for pair in device_pairs:
             self.recognised_hardware.add(pair)
+
+    def register_enumerate_func(self, func):
+        self.enumerate_func.add(func)
 
     def create_client(self, device, handler, plugin):
         # Get from cache first
@@ -504,6 +509,7 @@ class DeviceMgr(ThreadJob, PrintError):
         self.print_error("scanning devices...")
         with self.hid_lock:
             hid_list = hid.enumerate(0, 0)
+
         # First see what's connected that we know about
         devices = []
         for d in hid_list:
@@ -518,6 +524,10 @@ class DeviceMgr(ThreadJob, PrintError):
                 id_ += str(interface_number) + str(usage_page)
                 devices.append(Device(d['path'], interface_number,
                                       id_, product_key, usage_page))
+
+        # Let plugin handlers enumerate devices we don't know about
+        for f in self.enumerate_func:
+            devices.extend(f())
 
         # Now find out what was disconnected
         pairs = [(dev.path, dev.id_) for dev in devices]
